Move things in the git package
