--- src/gwt/src/org/rstudio/studio/client/workbench/views/source/editors/text/ace/AceDiffGutter.css
@@ -1,7 +1,0 @@
-.added {
-    border-right: 2px solid green;
-}
-
-.modified {
-    border-right: 2px solid blue;
-}
