+++ src/gwt/src/org/rstudio/studio/client/workbench/views/source/editors/text/ace/LineWidgetManager.java
@@ -22,7 +22,9 @@ public class LineWidgetManager extends JavaScriptObject
    protected LineWidgetManager() {}
    
    public native final void addLineWidget(LineWidget widget) /*-{
-      
+      console.log("addLineWidget");
+      console.log(widget);
+
       // avoid duplicating a pre-existing line widget
       var widgets = this.session.lineWidgets || {};
       if (widgets[widget.row] === widget)
